Return an int to match snprintf() prototype

+++ snfmt.h
@@ -28,14 +28,14 @@ union snfmt_arg {
 	void *p;
 };
 
-typedef size_t snfmt_func(char *, size_t, const char *, union snfmt_arg *);
+typedef int snfmt_func(char *, size_t, const char *, union snfmt_arg *);
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-size_t snfmt(snfmt_func *, char *, size_t, const char *, ...);
-size_t snfmt_va(snfmt_func *, char *, size_t, const char *, va_list);
+int snfmt(snfmt_func *, char *, size_t, const char *, ...);
+int snfmt_va(snfmt_func *, char *, size_t, const char *, va_list);
 
 #ifdef __cplusplus
 }
@@ -47,6 +47,6 @@ size_t snfmt_va(snfmt_func *, char *, size_t, const char *, va_list);
  * unreachable call to snprintf() on which MSVC performs the checks.
  */
 #define snfmt(cb, p, n, ...) \
-	(0 ? (size_t)snprintf((p), (n), ## __VA_ARGS__) : snfmt((cb), (p), (n), ## __VA_ARGS__))
+	(0 ? snprintf((p), (n), ## __VA_ARGS__) : snfmt((cb), (p), (n), ## __VA_ARGS__))
 
 #endif
